# osmpbf: preallocation node tags array