--- a/go.mod
+++ b/go.mod
@@ -6,6 +6,7 @@ require (
 	github.com/datadog/czlib v0.0.0-20160811164712-4bc9a24e37f2
 	github.com/gogo/protobuf v1.3.1
 	github.com/paulmach/orb v0.1.6
-	github.com/paulmach/protoscan v0.1.0
+	github.com/paulmach/protoscan v0.2.1
 	golang.org/x/time v0.0.0-20190921001708-c4c64cad1fd0
+	google.golang.org/genproto v0.0.0-20200526211855-cb27e3aa2013 // indirect
 )