also upgrade protoscan v0.1.0 -> v0.2.1

I don't know if I fully believe this benchmark, but looks pretty good

```
benchmark                        old ns/op     new ns/op     delta
BenchmarkLondon-12               261982560     247342371     -5.59%
BenchmarkLondon_nodes-12         202910043     183471063     -9.58%

benchmark                        old allocs     new allocs     delta
BenchmarkLondon-12               5145476        4804810        -6.62%
BenchmarkLondon_nodes-12         3732515        3391830        -9.13%

benchmark                        old bytes     new bytes     delta
BenchmarkLondon-12               951801475     918215396     -3.53%
BenchmarkLondon_nodes-12         646403798     612815672     -5.20%
```